+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added Catppuccin Latte theme
 - Added Dracula theme
 - Added Toky Night theme
+- Added Nord Light/Dark themes
 - Added Adwaita Light/Dark themes
 - Added Breeze Light/Dark themes